+++ src/CosmosDB/CosmosDB/ChangeLog.md
@@ -20,6 +20,7 @@
 
 ## Upcoming Release
 * Fixed issue where the restore command failed to detect the source account, even when it existed.
+* Added support for preserving existing throughput buckets during SQL container throughput updates.
 
 ## Version 2.0.7
 * Added support for throughput buckets.